+++ src/test/java/com/albion/common/graph/algorithms/BreathFirstSearchTest.java
@@ -1,6 +1,5 @@
 package com.albion.common.graph.algorithms;
 
-import com.albion.common.graph.algorithms.BreathFirstSearch;
 import com.albion.common.graph.core.v1.Graph;
 import com.albion.common.graph.core.v1.Vertex;
 import org.testng.Assert;
